Worker1/requirements.txt
python-dotenv
requests
schedule
coloredlogs
pandas
splunk-sdk